+++ tests/unit/__init__.py
@@ -1,1 +1,1 @@
-from tests.unit import dispatch, managers, viewsets, decorators, exceptions, tests_utilities
+from . import dispatch, managers, viewsets, decorators, exceptions, tests_utilities, tests