+++ CHANGELOG.rst
@@ -7,7 +7,7 @@ CHANGELOG
 - Endpoints that are generated now use "<class name>__<function_name>" as the default endpoint name
 - Removed check fore the base rest endpoint.  This doesn't take into account dispatchers having url_prefixes
 - Added get method to RequestContainer which searches through the url_params, query_args, and then the body_args to find a value
-
+- Added create_fields property to the BaseManager.  This allows a user to specify which fields are valid for creation.
 
 0.2.0 (2015-04-28)
 ==================